feat: 05-merge-styles OK2

+++ 04-copy-directory/index.js
@@ -7,8 +7,9 @@ const copyDir = async function () {
    const rmdir = await fsPromises.rm(path.resolve(__dirname, 'files-copy'), { recursive: true, force: true });
    const mkdir = await fsPromises.mkdir(path.resolve(__dirname, 'files-copy'), { recursive: true });
    for (const file of files) {
-      //console.log(file);
-      const copy = await fsPromises.copyFile(path.resolve(__dirname, 'files', file.name), path.resolve(__dirname, 'files-copy', file.name));
+      if (file.isFile()) {
+         const copy = await fsPromises.copyFile(path.resolve(__dirname, 'files', file.name), path.resolve(__dirname, 'files-copy', file.name));
+      }
    }
 }
